Clean the code
